chore(readme): add information about inspiration for plugin (#34)
This pull request includes an update to the `README.md` file to provide
additional context and inspiration for the IntelliJ IDEA plugin. The
most important change is the addition of a new section that explains the
inspiration behind the plugin and its intended improvements.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R10-R17):
Added an "Inspiration" section detailing the origins of the plugin, its
goals for flexibility and extensibility, and a call for feedback and
contributions.